BX-2499/tx-msg-with-double-ts (#186)

+++ lib/libtpe/tpe/service/transaction_service.cpp
@@ -388,8 +388,8 @@ TxProcessingResult_t TransactionService::process_transaction_msg(
     const Sha256_t& transaction_hash,
     PTxContents_t transaction_contents,
     unsigned int short_id,
-    unsigned int timestamp,
-    unsigned int current_time,
+    double timestamp,
+    double current_time,
     bool enable_transaction_validation,
     uint64_t min_tx_network_fee,
     bool from_relay
@@ -791,8 +791,8 @@ std::tuple<TxStatus_t , TxValidationStatus_t> TransactionService::_msg_tx_build_
     unsigned int short_id,
     const Sha256_t& transaction_hash,
     const PTxContents_t& transaction_contents,
-    unsigned int timestamp,
-    unsigned int current_time,
+    double timestamp,
+    double current_time,
     bool enable_transaction_validation,
     uint64_t min_tx_network_fee,
     bool from_relay
